refactor(billing): keep the managed model weights with the catalogue they price
The guard added a moment ago lives under packages/server/api/test/unit,
which no CI job runs, so it could not have stopped an unpriced model from
merging. Pointed out in review, and disclosing it in the description was
the weaker answer when the data can simply live somewhere CI reaches.

MANAGED_MODEL_WEIGHTS now sits in core-piece-types beside
ALLOWED_CHAT_MODELS_BY_PROVIDER and ACTIVEPIECES_CHAT_TIERS, which already
carries creditWeight for the three tiers, so all the pricing data is in one
place and the package is in the main CI test filter. resolveAiCreditWeight
still owns the lookup and reads the table through @activepieces/shared, and
the default rate is a named constant rather than a literal at the call
site.

The data guard moved with it, so adding an offered model without a weight
now fails a test CI runs. Two more assertions came out of the move: a
pending entry has to still be offered, so an exception cannot outlive the
model it excuses, and nothing may be both priced and pending. The api test
keeps only what exercises resolveAiCreditWeight itself.
