+++ packages/core/piece-types/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@activepieces/core-piece-types",
-  "version": "0.10.0",
+  "version": "0.11.0",
   "type": "commonjs",
   "main": "./dist/src/index.js",
   "scripts": {
