+++ packages/core/shared/src/lib/management/ai-providers/index.ts
@@ -455,6 +455,9 @@ export {
     AI_PROVIDER_ENTITY_TYPES,
     ALLOWED_CHAT_MODELS_BY_PROVIDER,
     ACTIVEPIECES_CHAT_TIERS,
+    MANAGED_MODEL_WEIGHTS,
+    MODELS_AWAITING_A_CREDIT_WEIGHT,
+    DEFAULT_MANAGED_MODEL_WEIGHT,
     DEFAULT_CHAT_TIER_ID,
     AI_PROVIDER_CAPABILITIES,
     OPENAI_COMPATIBLE_VENDOR_BASE_URLS,
